remove testLeaseRecovery from LEGACY
